requirements.txt
langchain
langchain-community
langchain-core
langgraph
langchain-groq
faiss-cpu
pypdf
fastapi
uvicorn
streamlit
python-dotenv
tiktoken
langchain-huggingface
sentence-transformers